add test, improve form

+++ src/components/Hit.js
@@ -21,31 +21,37 @@ const Hit = ({ hit }) => {
     }
   };
 
-  return (
-    <div>
-      {isDeleted ? (
-        <div>Restaurant is successfully removed!</div>
-      ) : (
-        <div>
-          {/* <img alt={hit.name} src={`${hit.image_url}`} /> */}
+  if (!hit) {
+    return <div>Loading...</div>;
+  }
+
+  if (hit) {
+    return (
+      <div>
+        {isDeleted ? (
+          <div>Restaurant is successfully removed!</div>
+        ) : (
           <div>
+            {/* <img alt={hit.name} src={`${hit.image_url}`} /> */}
             <div>
-              <Highlight attribute='name' hit={hit} />
+              <div data-testid='restaurant-name'>
+                <Highlight attribute='name' hit={hit} />
+              </div>
+              <div>Price range is {hit.price_range}</div>
+              <div>Cousine: {hit.food_type}</div>
+              <div>Rating: {hit.stars_count}</div>
+              <button
+                className='rounded border border-gray-400 px-2 py-1 mt-2'
+                onClick={() => deleteItem(hit.objectID)}
+              >
+                Remove
+              </button>
             </div>
-            <div>Price range is {hit.price_range}</div>
-            <div>Cousine: {hit.food_type}</div>
-            <div>Rating: {hit.stars_count}</div>
-            <button
-              className='rounded border border-gray-400 px-2 py-1'
-              onClick={() => deleteItem(hit.objectID)}
-            >
-              Delete
-            </button>
           </div>
-        </div>
-      )}
-    </div>
-  );
+        )}
+      </div>
+    );
+  }
 };
 
 export default Hit;
